WIPSY TIPSY

+++ app_runner/docker_app_runner/app_runner_test.go
@@ -45,7 +45,7 @@ var _ = Describe("AppRunner", func() {
 				Instances:            22,
 				MemoryMB:             128,
 				DiskMB:               1024,
-				Ports:                docker_app_runner.PortConfig{Exposed: []uint32{2000, 4000}, Monitored: 2000},
+				Ports:                docker_app_runner.PortConfig{Exposed: []uint16{2000, 4000}, Monitored: 2000},
 				WorkingDir:           "/user/web/myappdir",
 			})
 			Expect(err).To(BeNil())
@@ -65,7 +65,7 @@ var _ = Describe("AppRunner", func() {
 				Routes:               []string{"americano-app.myDiegoInstall.com"},
 				MemoryMB:             128,
 				DiskMB:               1024,
-				Ports:                []uint32{2000, 4000},
+				Ports:                []uint16{2000, 4000},
 				LogGuid:              "americano-app",
 				LogSource:            "APP",
 				Setup: &models.DownloadAction{
@@ -103,7 +103,7 @@ var _ = Describe("AppRunner", func() {
 						MemoryMB:             128,
 						Monitor:              true,
 						DiskMB:               1024,
-						Ports:                docker_app_runner.PortConfig{Exposed: []uint32{8080}, Monitored: 6682},
+						Ports:                docker_app_runner.PortConfig{Exposed: []uint16{8080}, Monitored: 6682},
 					})
 
 					Expect(err).To(HaveOccurred())
@@ -122,7 +122,7 @@ var _ = Describe("AppRunner", func() {
 					DockerImagePath: "runtest/runner",
 					AppArgs:         []string{},
 					Monitor:         false,
-					Ports:           docker_app_runner.PortConfig{Monitored: 1234, Exposed: []uint32{1234}},
+					Ports:           docker_app_runner.PortConfig{Monitored: 1234, Exposed: []uint16{1234}},
 				})
 				Expect(err).ToNot(HaveOccurred())
 
@@ -145,7 +145,7 @@ var _ = Describe("AppRunner", func() {
 				Instances:            1,
 				MemoryMB:             128,
 				DiskMB:               1024,
-				Ports:                docker_app_runner.PortConfig{Monitored: 8080, Exposed: []uint32{8080}},
+				Ports:                docker_app_runner.PortConfig{Monitored: 8080, Exposed: []uint16{8080}},
 			})
 
 			Expect(err).To(HaveOccurred())
@@ -165,7 +165,7 @@ var _ = Describe("AppRunner", func() {
 					Instances:            1,
 					MemoryMB:             128,
 					DiskMB:               1024,
-					Ports:                docker_app_runner.PortConfig{Exposed: []uint32{8080}, Monitored: 8080},
+					Ports:                docker_app_runner.PortConfig{Exposed: []uint16{8080}, Monitored: 8080},
 				})
 
 				Expect(err).To(HaveOccurred())
@@ -188,7 +188,7 @@ var _ = Describe("AppRunner", func() {
 					Instances:            1,
 					MemoryMB:             128,
 					DiskMB:               1024,
-					Ports:                docker_app_runner.PortConfig{Exposed: []uint32{8080}, Monitored: 8080},
+					Ports:                docker_app_runner.PortConfig{Exposed: []uint16{8080}, Monitored: 8080},
 				})
 
 				Expect(err).To(Equal(upsertError))
@@ -209,7 +209,7 @@ var _ = Describe("AppRunner", func() {
 					Instances:            1,
 					MemoryMB:             128,
 					DiskMB:               1024,
-					Ports:                docker_app_runner.PortConfig{Exposed: []uint32{8080}, Monitored: 8080},
+					Ports:                docker_app_runner.PortConfig{Exposed: []uint16{8080}, Monitored: 8080},
 				})
 
 				Expect(err).To(Equal(receptorError))
@@ -229,7 +229,7 @@ var _ = Describe("AppRunner", func() {
 					Instances:            1,
 					MemoryMB:             128,
 					DiskMB:               1024,
-					Ports:                docker_app_runner.PortConfig{Exposed: []uint32{8080}, Monitored: 8080},
+					Ports:                docker_app_runner.PortConfig{Exposed: []uint16{8080}, Monitored: 8080},
 				})
 
 				Expect(err).To(Equal(receptorError))
